* archive.c (_bfd_archive_bsd_update_armap_timestamp): Cast stat 	st_mtime to long before comparison.

+++ bfd/archive.c
@@ -2311,7 +2311,7 @@ _bfd_archive_bsd_update_armap_timestamp (bfd *arch)
       /* Can't read mod time for some reason.  */
       return TRUE;
     }
-  if (archstat.st_mtime <= bfd_ardata (arch)->armap_timestamp)
+  if (((long) archstat.st_mtime) <= bfd_ardata (arch)->armap_timestamp)
     /* OK by the linker's rules.  */
     return TRUE;
